AdvertDesk/requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
astroid==3.3.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.0
dill==0.3.8
Django==5.1
django-allauth==64.0.0
django-apscheduler==0.6.2
django-ckeditor-5==0.2.13
django-filter==24.3
django-js-asset==2.2.0
django-modeltranslation==0.19.6
django-mptt==0.16.0
djangorestframework==3.15.2
idna==3.7
isort==5.13.2
jwt==1.3.1
kombu==5.4.0
Markdown==3.6
mccabe==0.7.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pycparser==2.22
pylint==3.2.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytils==0.4.1
pytz==2024.1
redis==5.0.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
tomlkit==0.13.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
